[#2822] Move trekking cirkwi in cirkwi app
